draw game inplemented && styles fixed

+++ src/components/Game/Game.js
@@ -9,10 +9,11 @@ class Game extends Component {
     super(props);
     this.state = {
       field: props.field || [...Array(props.sizeY)].map(() => [...Array(props.sizeX)]),
-      cross: props.cross || true,
-      win: false,
-      matchToWin: props.matchToWin || 5,
-      escPressed: false,
+      isCrossTurn: props.isCrossTurn,
+      isGameWon: false,
+      isGameDrawn: false,
+      isMenuShowed: false,
+      matchToWin: props.matchToWin,
     };
   }
 
@@ -28,7 +29,7 @@ class Game extends Component {
     const { matchToWin } = this.state;
     let inRow = 1;
 
-    array.reduce((a, b, _, currentReducedArray) => {
+    array.slice().reduce((a, b, _, currentReducedArray) => {
       if (a && a === b) inRow += 1;
       else inRow = 1;
       if (inRow === matchToWin) currentReducedArray.splice(1);
@@ -36,6 +37,8 @@ class Game extends Component {
     });
     return inRow === matchToWin;
   };
+
+  checkDraw = field => field.map(row => row.filter(x => !x)).filter(y => y.length !== 0).length === 0;
 
   getDiagonale = (field, rowIndex, cellIndex, isMain) => {
     const { matchToWin } = this.state;
@@ -64,22 +67,21 @@ class Game extends Component {
     const { state } = this;
     const { sizeX, sizeY } = this.props;
 
-    if (!state.field[rowIndex][cellIndex]) {
+    if (!state.field[rowIndex][cellIndex] && !state.isGameWon && !state.isGameDrawn) {
       this.setState(prevState => {
-        const { field, cross } = prevState;
+        const { field, isCrossTurn } = prevState;
 
-        field[rowIndex][cellIndex] = cross ? 'x' : 'o';
-        const win = this.winCondition(field, rowIndex, cellIndex);
-        if (!win) {
+        field[rowIndex][cellIndex] = isCrossTurn ? 'x' : 'o';
+        const isGameWon = this.winCondition(field, rowIndex, cellIndex);
+        const isGameDrawn = this.checkDraw(field);
+        if (!isGameWon && !isGameDrawn) {
           localStorage.setItem('field', JSON.stringify(field));
           localStorage.setItem('sizeX', sizeX);
           localStorage.setItem('sizeY', sizeY);
           localStorage.setItem('matchToWin', state.matchToWin);
-          console.log(!cross);
-          localStorage.setItem('cross', !cross);
+          localStorage.setItem('isCrossTurn', !isCrossTurn);
         } else localStorage.clear();
-
-        return { field, cross: !cross, win };
+        return { field, isCrossTurn: !isCrossTurn, isGameWon, isGameDrawn };
       });
     }
   };
@@ -89,55 +91,57 @@ class Game extends Component {
     localStorage.clear();
     this.setState({
       field: [...Array(sizeY)].map(() => [...Array(sizeX)]),
-      cross: true,
-      win: false,
-      escPressed: false,
+      isCrossTurn: true,
+      isGameWon: false,
+      isGameDrawn: false,
+      isMenuShowed: false,
     });
   };
 
   handleEscKey = e => {
     if (e.keyCode === 27) {
       this.setState(prevState => ({
-        escPressed: !prevState.escPressed,
+        isMenuShowed: !prevState.isMenuShowed,
       }));
     }
   };
 
   render() {
     const { sizeX, sizeY, quit } = this.props;
-    const { field, cross, win, escPressed } = this.state;
-    const cellSize = 40;
-    if (win) return <h1>{`Congrats, ${cross ? 'circle' : 'cross'}!`}</h1>;
+    const { field, isCrossTurn, isGameWon, isGameDrawn, isMenuShowed } = this.state;
+    const CELL_SIZE = 40;
+
     return (
       <div className="game">
-        <Stage width={sizeX * cellSize + 2} height={sizeY * cellSize + 2}>
+        <p>Press Esc to open Menu</p>
+        <Stage width={sizeX * CELL_SIZE + 2} height={sizeY * CELL_SIZE + 2}>
           <Layer>
-            <Rect x={0} y={0} width={sizeX * cellSize + 2} height={sizeY * cellSize + 2} fill="#bfbfbf" />
+            <Rect x={0} y={0} width={sizeX * CELL_SIZE + 2} height={sizeY * CELL_SIZE + 2} fill="#bfbfbf" />
             {field.map((row, rowIndex) =>
               row.map((cell, cellIndex) => (
                 <React.Fragment key={`canvas-row-${rowIndex + 1}-cell-${cellIndex + 1}`}>
                   <Rect
-                    x={cellIndex * cellSize + 1}
-                    y={rowIndex * cellSize + 1}
-                    width={cellSize}
-                    height={cellSize}
+                    x={cellIndex * CELL_SIZE + 1}
+                    y={rowIndex * CELL_SIZE + 1}
+                    width={CELL_SIZE}
+                    height={CELL_SIZE}
                     fill="#bfbfbf"
                   />
                   <Rect
-                    x={cellIndex * cellSize + 2}
-                    y={rowIndex * cellSize + 2}
-                    width={cellSize - 2}
-                    height={cellSize - 2}
+                    x={cellIndex * CELL_SIZE + 2}
+                    y={rowIndex * CELL_SIZE + 2}
+                    width={CELL_SIZE - 2}
+                    height={CELL_SIZE - 2}
                     fill="#ebe8be"
                     onClick={() => this.handleClick(rowIndex, cellIndex)}
                   />
                   <Text
-                    x={cellIndex * cellSize}
-                    y={rowIndex * cellSize}
+                    x={cellIndex * CELL_SIZE}
+                    y={rowIndex * CELL_SIZE}
                     text={cell}
                     fontStyle="bold"
-                    offsetX={-cellSize / 5}
-                    fontSize={cellSize}
+                    offsetX={-CELL_SIZE / 5}
+                    fontSize={CELL_SIZE}
                     fill={cell === 'x' ? '#bf866b' : '#c2c186'}
                   />
                 </React.Fragment>
@@ -146,8 +150,14 @@ class Game extends Component {
           </Layer>
         </Stage>
 
-        {escPressed && (
+        {(isMenuShowed || isGameWon || isGameDrawn) && (
           <div className="menu">
+            <h1>
+              {isGameWon || isGameDrawn
+                ? (isGameWon && `Congrats, ${isCrossTurn ? 'circle' : 'cross'}!`) ||
+                  (isGameDrawn && 'Ooops! This is a draw!')
+                : 'Menu'}
+            </h1>
             <button type="button" className="btn" onClick={this.restart}>
               Restart
             </button>
@@ -166,7 +176,7 @@ Game.propTypes = {
   sizeX: PropTypes.number,
   sizeY: PropTypes.number,
   quit: PropTypes.func.isRequired,
-  cross: PropTypes.bool,
+  isCrossTurn: PropTypes.bool,
   matchToWin: PropTypes.number,
 };
 
@@ -174,7 +184,7 @@ Game.defaultProps = {
   field: undefined,
   sizeX: 10,
   sizeY: 10,
-  cross: true,
+  isCrossTurn: true,
   matchToWin: 5,
 };
